requirements.txt
Flask
SQLAlchemy
faiss-cpu
torch
transformers
python-dotenv
requests
PyJWT
werkzeug
email-validator
Pillow
Flask-SQLAlchemy